% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{al_simple_media_err}
\alias{al_simple_media_err}
\title{Muestreo aleatorio simple con error absoluto}
\usage{
al_simple_media_err(s2, e, N)
}
\arguments{
\item{s2}{Varianza muestral}

\item{e}{error absoluto}

\item{N}{error absoluto}
}
\value{
Número de muestra óptimo.
}
\description{
Función que obtiene el tamaño de muestra,basado en la media, utilizando muestreo aleatorio simple.
}
\examples{
Tamaño de la población
al_simple_media_err(s2=3,e=5,50);
}
