% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{al_simple_prop_err_rel}
\alias{al_simple_prop_err_rel}
\title{Muestreo aleatorio simple con error relativo, basado en proporción polacional.}
\usage{
al_simple_prop_err_rel(p, e, N)
}
\arguments{
\item{p}{proporción con característica de interés}

\item{e}{Error relativo}

\item{N}{Tamaño de la población}
}
\value{
Número de muestra óptimo
}
\description{
Función que obtiene el tamaño de muestra, basado en proporción polacional, utilizando muestreo aleatorio simple.
}
\examples{
al_simple_prop_err_rel(p=0.3,e=0.05,N=50);
}
