% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{al_simple_total_err}
\alias{al_simple_total_err}
\title{Muestreo aleatorio simple con error absoluto, basado en el total de la población.}
\usage{
al_simple_total_err(s2, e, N)
}
\arguments{
\item{s2}{Varianza muestral}

\item{e}{Error absoluto}

\item{N}{Tamaño de la población}
}
\value{
Número de muestra óptimo
}
\description{
Función que obtiene el tamaño de muestra, basado en el total de la población, utilizando muestreo aleatorio simple.
}
\examples{
al_simple_total_err(s2=3,e=5,N=50, u=10);
}
