% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{al_simple_total_err_rel}
\alias{al_simple_total_err_rel}
\title{Muestreo aleatorio simple con error relativo, basado en el total de la población.}
\usage{
al_simple_total_err_rel(s2, e, N, u)
}
\arguments{
\item{s2}{Varianza muestral}

\item{e}{Error relativo}

\item{N}{Tamaño de la población}

\item{u}{Media poblacional}
}
\value{
Número de muestra óptimo
}
\description{
Función que obtiene el tamaño de muestra, basado en el total de la población, utilizando muestreo aleatorio simple.
}
\examples{
al_simple_total_err_rel(s2=3,e=0.05,N=50, u=10);
}
